multi ==> dual

+++ include/CSV.h
@@ -7,5 +7,5 @@ namespace CSV {
 auto loadImages(const std::string &filename) noexcept
     -> std::vector<std::vector<double>>;
 
-auto loadLabels(const std::string &filename) noexcept -> std::vector<double>;
+auto loadLabels(const std::string &filename) noexcept -> std::vector<int>;
 }  // namespace CSV
